Keep CTF directory names in JSON config file
Because the location of these files is likely to differ between environments but
not between sessions in the same environment, I feel the best solution is to
ask the user to choose them if they haven't already, and then store their
choices for future.

It *may* be preferable to have a config file per user, but i've avoided this for
now, based on trepidation about OS differences in user vs shared areas.

This commit just looks for the value in the config file and falls back on a
default, 'T:\CMIS\CTF Files\' if the value is not found.

And for the time-being I've only stored the name of the parent folder, retaining
the assumption that the names of the child folders will remain:
CTF_In
CTF_Year07_2020_2021
CTF_Year12_2020_2021

Future work should remove this assumption.

Added .vscode dir to gitignore. It was created to be the home of the debug
launcher.