Begin creating test suite
Decided to go with pytest over unittest and nose. pytest seems to be the more
active project and preferred in the community. Also, nose's use of custom assert
syntax suggests it would be harder to move away from in the future.

I have used pipenv to manage dependencies - a problem pulled into focus now
that pytest is being used (since it's not in the standard library). I added
pylint to pipenv, but still can't seem to get it to work seemlessly in Winows.

On windows, I have to run tests using `pipenv run py -m pytest`
This ought to work: `pipenv run pytest`
But path problems with pytest are preventing it at the moment.

I think tests will be more managable split into multiple files, even though
they'll all test parse_CTFs.py. Something like this division of labour:
* case fixing
* empty node trimming
* xml checking
* file management?
* config management
* end-to-end tests

I've added some example CTFs in test/CTF, to be used in end-to-end tests