Add installation instructions to README
I've tested theses instructions on Windows and MacOS (using Homebrew and Bash).
I've assumed the MacOS process applies to other Unix-like OSs based on the
documentation I've seen.

On Unix systems, if pipenv determines that the required version of Python is not
installed, it will look for pyenv and if it finds it, uses it to install the
Python version. I haven't confirmed that on Windows pipenv will attempt the same
thing with pyenv-win (the Windows port of pyenv).