Make sure Surname is the same as Legal Surname
Decided to raise a ValueError if a pupil is missing both their Surname and their
PreferredSurname, because this inevitably means the import will fail. If an
error is raised, the UPNs are reported to the user.
Of course, even if one student has dodgy data, you could perhaps import the
others in the file, so if this occurs with any regularity it might be best to
demote this to a warning.
For consistency I changed the ac_year function to raise ValueError (instead of
NameError, which was inappropriate). Ultimately a custom error class will be
the correct solution.