Cleanup case-fixing tests
Removed the utility function, add_xml_nodes_from_dict in the main file -- it was
never called in the main code. I decided this was speculative generality/
functionality, and it would be safer to rely solely on the library, dicttoxml,
which I'd already used in other tests.

I also added a nested element that needs to be proper-cased. To maintain
readibility I changed the fixtures so that the before/after values are in 2
seperate dicts originally.